requirements.txt
streamlit
langchain==0.1.0
langchain-community==0.0.20
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.16.4
faiss-cpu
PyPDF2
python-dotenv